requirements.txt
torch==2.5.1
torchvision==0.20.1
transformers==4.48.2
pillow==11.1.0
opencv-python==4.11.0.86
streamlit==1.45.0
gtts==2.3.1
speechrecognition==3.10.0
accelerate>=0.26.0
playsound==1.2.2
pyaudio==0.2.13